+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unpublished]
+### Added
+- Added `)userinfo` command
+
+### Removed
+- Removed `~id` command
+
 ## [2.3.2] - 2016-07-08
 ### Fixed
 - Fixed `~alog` command from showing '&amp;apos&semi;' for K'ril and Kree kills.